feat(utility): Add ssize function
Add a new `ssize` function to the `stdcpp::` namespace in the `utility.hpp` file. This function is used to return the signed size of a container, by casting its size to a signed type. It takes advantage of perfect forwarding and uses std::forward<T> for efficiency.

The purpose of this addition is to provide an easy way to obtain the signed size of any container, which can be useful in various scenarios where negative sizes are needed or preferred.

This commit also includes an additional include statement for `<utility>` as it is required for using std::forward<T>.